requirements.txt
PyYAML
selenium
beautifulsoup4
webdriver_manager
pydantic
